Обновление лабораторной #4
Обновление лабораторной работы, чтобы была возмоность запуска с помощью Maven в соостветсвии с быстрым стартом. И также небольшие испралвения.
