--- lab04/source/CompareCollection/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=8aaa431e
-build.xml.script.CRC32=a92d0605
-build.xml.stylesheet.CRC32=f85dc8f2@1.94.0.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=8aaa431e
-nbproject/build-impl.xml.script.CRC32=c368badb
-nbproject/build-impl.xml.stylesheet.CRC32=f89f7d21@1.94.0.48
