--- lab04/source/CompareCollection/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\lenovo\\AppData\\Roaming\\NetBeans\\11.3\\build.properties
