--- lab04/source/CompareCollection/nbproject/private/private.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/G:/Java/CompareCollection/src/com/coordsystem/MainClass.java</file>
-            <file>file:/G:/Java/CompareCollection/src/com/coordsystem/primitives/Point.java</file>
-            <file>file:/G:/Java/CompareCollection/src/com/coordsystem/comparators/PointComparator.java</file>
-        </group>
-    </open-files>
-</project-private>
